UV_BOT/requirements.txt
python-telegram-bot==22.3
python-docx==1.2.0
pytz==2025.2